fix(agent): make agent integration lane non-vacuous
Address lalalune CR on #17838:

- Rename views-registry-integration.test.ts -> views-registry.integration.test.ts
  so the packages/agent/src/**/*.integration.test.ts glob discovers a real suite
- Drop --passWithNoTests from test:integration so empty collection fails
- Remove hand-enumeration of the suite from test.yml Views system tests and
  packages/agent test:remote-capabilities
- Fold glob-coverage assertions into the existing agent subpath-alias test
  and delete the near-duplicate globs file

Verified: bun run --cwd packages/agent test:integration -> 37 passed
