+++ packages/agent/src/__tests__/views-registry.integration.test.ts
